database/requirements.txt
psycopg2-binary
pymongo
neo4j
pyyaml